[ADF-4514] Move widgets pages to Testing package
