+++ lib/testing/src/lib/core/pages/form/widgets/dropdownWidget.ts
@@ -17,7 +17,7 @@
 
 import { FormFields } from '../formFields';
 import { by, element } from 'protractor';
-import { BrowserVisibility, BrowserActions } from '@alfresco/adf-testing';
+import { BrowserVisibility, BrowserActions } from '../../../utils/public-api';
 
 export class DropdownWidget {
